by manish 12-12-2023

+++ lib/admin/addVendor_screen.dart
@@ -72,7 +72,7 @@ class _AddVendorScreenState extends State<AddVendorScreen> {
         } else {
           imageUrl = fileUrl;
         }
-      } else {
+      }  else if (resturentData != null) { // Check if resturentData is not null
         if (!categoryFile.path.contains("https")) {
           Reference gg = FirebaseStorage.instance.refFromURL(resturentData!.image.toString());
           await gg.delete();
